+++ src/API/CourseProviderInterface.php
@@ -5,6 +5,7 @@ declare(strict_types=1);
 namespace Dbp\Relay\BaseCourseBundle\API;
 
 use Dbp\Relay\BaseCourseBundle\Entity\Course;
+use Dbp\Relay\BaseCourseBundle\Entity\CourseClass;
 use Dbp\Relay\CoreBundle\Exception\ApiError;
 
 interface CourseProviderInterface
@@ -29,4 +30,16 @@ interface CourseProviderInterface
      * @throws ApiError
      */
     public function getCourses(int $currentPageNumber, int $maxNumItemsPerPage, array $options = []): array;
+
+    public function getCourseClassById(string $identifier, array $options = []): ?CourseClass;
+
+    /**
+     * @param array $options Available options:
+     *                       * LocalData::INCLUDE_PARAMETER_NAME
+     *
+     * @return CourseClass[]
+     *
+     * @throws ApiError
+     */
+    public function getCourseClasses(int $currentPageNumber, int $maxNumItemsPerPage, array $options = []): array;
 }