+++ tests/ApiTest.php
@@ -9,21 +9,45 @@ use Symfony\Component\HttpFoundation\Response;
 
 class ApiTest extends AbstractApiTest
 {
-    public function testGetItemAuthenticated(): void
+    public function testGetCourseItemAuthenticated(): void
     {
         $this->testClient->setUpUser(userAttributes: ['MAY_READ' => true]);
         $this->assertEquals(Response::HTTP_OK, $this->testClient->get('/base/courses/foo')->getStatusCode());
     }
 
-    public function testGetItemNotFound(): void
+    public function testGetCourseCollectionAuthenticated(): void
+    {
+        $this->testClient->setUpUser(userAttributes: ['MAY_READ' => true]);
+        $this->assertEquals(Response::HTTP_OK, $this->testClient->get('/base/courses')->getStatusCode());
+    }
+
+    public function testGetCourseItemNotFound(): void
     {
         $this->testClient->setUpUser(userAttributes: ['MAY_READ' => true]);
         $this->assertEquals(Response::HTTP_NOT_FOUND, $this->testClient->get('/base/courses/404')->getStatusCode());
     }
 
+    public function testGetCourseDateItemAuthenticated(): void
+    {
+        $this->testClient->setUpUser(userAttributes: ['MAY_READ' => true]);
+        $this->assertEquals(Response::HTTP_OK, $this->testClient->get('/base/course-dates/bar')->getStatusCode());
+    }
+
+    public function testGetCourseDateCollectionAuthenticated(): void
+    {
+        $this->testClient->setUpUser(userAttributes: ['MAY_READ' => true]);
+        $this->assertEquals(Response::HTTP_OK, $this->testClient->get('/base/course-dates')->getStatusCode());
+    }
+
+    public function testGetCourseDateItemNotFound(): void
+    {
+        $this->testClient->setUpUser(userAttributes: ['MAY_READ' => true]);
+        $this->assertEquals(Response::HTTP_NOT_FOUND, $this->testClient->get('/base/course-dates/404')->getStatusCode());
+    }
+
     public function testUnauthorizedRequests()
     {
-        foreach (['/base/courses', '/base/courses/foo'] as $path) {
+        foreach ($this->getPaths() as $path) {
             $response = $this->testClient->get($path, token: null);
             $this->assertEquals(Response::HTTP_UNAUTHORIZED, $response->getStatusCode());
         }
@@ -31,10 +55,23 @@ class ApiTest extends AbstractApiTest
 
     public function testForbiddenRequests()
     {
-        foreach (['/base/courses', '/base/courses/foo'] as $path) {
+        foreach ($this->getPaths() as $path) {
             $this->testClient->setUpUser(userAttributes: ['MAY_READ' => false]);
             $response = $this->testClient->get($path);
             $this->assertEquals(Response::HTTP_FORBIDDEN, $response->getStatusCode());
         }
+    }
+
+    /**
+     * @return string[]
+     */
+    private function getPaths(): array
+    {
+        return [
+            '/base/courses',
+            '/base/courses/foo',
+            '/base/course-dates',
+            '/base/course-dates/bar',
+        ];
     }
 }